Sort existing tests into topic files
